Add custom class in between Click and our commands
This allows us to add custom logic, such as dynamically generating extra
commands which will be needed for the LGTVRemote commands.
